omega mobile ux fixes
