+++ components/blocks/FeatureCarousel/CarouselFeature.tsx
@@ -1,3 +1,4 @@
+import checkTouchScreen from 'components/util/touchscreenDetection';
 import Image from 'next/image';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { tinaField } from 'tinacms/dist/react';
@@ -7,27 +8,6 @@ import { icons } from '../../ui/IconPickerIcons';
 import { Actions } from '../ActionButton/ActionsButton';
 import { Container } from '../Container';
 import { CarouselFeatureMobile } from './CarouselFeature.mobile';
-
-const checkTouchScreen = () => {
-  let hasTouchScreen = false;
-  if ('maxTouchPoints' in navigator) {
-    hasTouchScreen = navigator.maxTouchPoints > 0;
-  } else {
-    const mQ = matchMedia?.('(pointer:coarse)');
-    if (mQ?.media === '(pointer:coarse)') {
-      hasTouchScreen = !!mQ.matches;
-    } else if ('orientation' in window) {
-      hasTouchScreen = true; // deprecated, but good fallback
-    } else {
-      // Only as a last resort, fall back to user agent sniffing
-      const UA: string = (navigator as Navigator).userAgent;
-      hasTouchScreen =
-        /\b(BlackBerry|webOS|iPhone|IEMobile)\b/i.test(UA) ||
-        /\b(Android|Windows Phone|iPad|iPod)\b/i.test(UA);
-    }
-  }
-  return hasTouchScreen;
-};
 
 const CarouselItem = ({
   data,
